Remove optional parameter handling from FlowClassConstructorParametersExtractor
Handling optional parameters (ones that have default values) is pointless,
because in the end we are going to invoke the flow providing only a list
of parameters. So they have to match constructor parameters list.

Allowing optional parameters this would only work if these were last
in the constructor parameters list. Otherwise, all parameters would be
moved one position backwards.
The above could lead to dangerous errors, especially if all parameters
are of the same type and the last one is optional too. The flow would
be started, but values would be assigned to wrong fields.

Therefore, for this part to work correctly, we enforce the client
to provide values for ALL parameters.