Change update method in FlowResultRepository semantics
Query now filters not only for clientId, but also for
null result field and RUNNING status.

This is to ensure given document is not updated multiple
times, potentially overriding previous values.
It is also important when considering multi-service
environment, especially after service restart - it is
to be implemented, but FlowExecutor should re-attach
to the node to wait for flows' results.
This change is a preparation for the above and an
improvement in general - it is safer to have
idempotent updates.