Update flowId inside update method in FlowResultRepository
This is another step to make the flow result process
more resilient - in case of an error during flowId
update earlier (FlowExecutor.startFlow method), flowId
is still going to be stored.
In case there was no error flowId field is updated
to the same value as before, which means no change
is made.