Change Integration Tests to mock at Controllers layer
Decided to make these tests almost Unit, because the
main purpose for their existence is to test Json
conversions and http responses. Because of that,
there is no real need to have the whole application
running. It became more obvious after introducing
MongoDB connection and realising that now all these
tests require MongoDB to run.

To overcome this issue, I decided to leave only
the routing part "real" and mock all Controllers.
This way the rest of the application is never going
to be created, because it will not be required.

Integration Tests that check the whole application
may still be introduced in the future.